Trait axum_server::service::MakeServiceRef

pub trait MakeServiceRef<Target, Request>: Sealed<(Target, Request)> {
    type Service: Service<Request, Response = Response<Self::Body>, Error = Self::Error, Future = Self::Future> + Send + 'static;
    type Body: Body<Data = Self::Data, Error = Self::BodyError> + Send + 'static;
    type Data: Send + 'static;
    type BodyError: Into<Box<dyn Error + Send + Sync>>;
    type Error: Into<Box<dyn Error + Send + Sync>>;
    type Future: Future<Output = Result<Response<Self::Body>, Self::Error>> + Send + 'static;
    type MakeError: Into<Box<dyn Error + Send + Sync>>;
    type MakeFuture: Future<Output = Result<Self::Service, Self::MakeError>>;
    fn poll_ready(
        &mut self, 
        cx: &mut Context<'_>
    ) -> Poll<Result<(), Self::MakeError>>;

    fn make_service(&mut self, target: &Target) -> Self::MakeFuture;
}

Modified version of MakeService that has some trait bounds and takes a &Target.

This trait is sealed and cannot be implemented for types outside this crate.

Associated Types

type Service: Service<Request, Response = Response<Self::Body>, Error = Self::Error, Future = Self::Future> + Send + 'static

type Body: Body<Data = Self::Data, Error = Self::BodyError> + Send + 'static

type Data: Send + 'static

type BodyError: Into<Box<dyn Error + Send + Sync>>

type Error: Into<Box<dyn Error + Send + Sync>>

type Future: Future<Output = Result<Response<Self::Body>, Self::Error>> + Send + 'static

type MakeError: Into<Box<dyn Error + Send + Sync>>

type MakeFuture: Future<Output = Result<Self::Service, Self::MakeError>>

Required methods

fn poll_ready(
&mut self,
cx: &mut Context<'_>
) -> Poll<Result<(), Self::MakeError>>

fn make_service(&mut self, target: &Target) -> Self::MakeFuture