Trait axum_server::service::SendService

pub trait SendService<Request>: Sealed<Request> {
    type Service: Service<Request, Response = Response<Self::Body>, Error = Self::Error, Future = Self::Future> + Send + 'static;
    type Body: Body<Data = Self::BodyData, Error = Self::BodyError> + Send + 'static;
    type BodyData: Send + 'static;
    type BodyError: Into<Box<dyn Error + Send + Sync>>;
    type Error: Into<Box<dyn Error + Send + Sync>>;
    type Future: Future<Output = Result<Response<Self::Body>, Self::Error>> + Send + 'static;
    fn into_service(self) -> Self::Service;
}

Trait alias for Service with bounds required for serve.

This trait is sealed and cannot be implemented for types outside this crate.

Associated Types

type Service: Service<Request, Response = Response<Self::Body>, Error = Self::Error, Future = Self::Future> + Send + 'static

type Body: Body<Data = Self::BodyData, Error = Self::BodyError> + Send + 'static

type BodyData: Send + 'static

type BodyError: Into<Box<dyn Error + Send + Sync>>

type Error: Into<Box<dyn Error + Send + Sync>>

type Future: Future<Output = Result<Response<Self::Body>, Self::Error>> + Send + 'static

Required methods

fn into_service(self) -> Self::Service