Struct axum_server::tls::TlsServer

pub struct TlsServer { /* fields omitted */ }

This is supported on crate feature tls-rustls only.

Configurable HTTP and HTTPS server, supporting HTTP/1.1 and HTTP2.

See module page for examples.

Implementations

impl TlsServer

pub fn new() -> Self

Create a new TlsServer.

pub fn bind<A, I>(self, addr: A) -> Self where
A: ToSocketAddrs<Iter = I> + Send + 'static,
I: Iterator<Item = SocketAddr> + 'static,

Bind to a single address or multiple addresses.

pub fn handle(self, handle: Handle) -> Self

Provide a Handle.

Successive calls will overwrite last Handle.

pub fn bind_rustls<A, I>(self, addr: A) -> Self where
A: ToSocketAddrs<Iter = I> + Send + 'static,
I: Iterator<Item = SocketAddr> + 'static,

Bind to a single address or multiple addresses. Using tls protocol on streams.

Certificate and private key must be set before or after calling this.

pub fn loader(self, loader: TlsLoader) -> Self

Provide a loaded TlsLoader.

This will overwrite any previously set private key and certificate(s) with its own ones.

pub fn tls_config(self, config: Arc<ServerConfig>) -> Self

Provide ServerConfig containing private key and certificate(s).

When this value is set, other tls configurations are ignored.

Successive calls will overwrite last value.

pub fn private_key(self, private_key: Vec<u8>) -> Self

Set private key in PEM format.

Successive calls will overwrite last private key.

pub fn certificate(self, certificate: Vec<u8>) -> Self

Set certificate(s) in PEM format.

Successive calls will overwrite last certificate.

pub fn private_key_file(self, path: impl AsRef<Path>) -> Self

Set private key from file in PEM format.

Successive calls will overwrite last private key.

pub fn certificate_file(self, path: impl AsRef<Path>) -> Self

Set certificate(s) from file in PEM format.

Successive calls will overwrite last certificate.

pub async fn serve<S, B>(self, service: S) -> Result<()> where
S: Service<Request<Body>, Response = Response<B>> + Send + Sync + 'static + Clone,
S::Error: Into<Box<dyn Error + Send + Sync>>,
S::Future: Send,
B: Body + Send + 'static,
B::Data: Send,
B::Error: Into<Box<dyn Error + Send + Sync>>,

Serve provided cloneable service on all binded addresses.

If accepting connection fails in any one of binded addresses, listening in all binded addresses will be stopped and then an error will be returned.

pub async fn serve_and_record<S, B>(self, service: S) -> Result<()> where
S: Service<Request<Body>, Response = Response<B>> + Send + Sync + 'static + Clone,
S::Error: Into<Box<dyn Error + Send + Sync>>,
S::Future: Send,
B: Body + Send + 'static,
B::Data: Send,
B::Error: Into<Box<dyn Error + Send + Sync>>,

This is supported on crate feature record only.

Serve provided cloneable service on all binded addresses.

Record sent and received bytes for each connection independently. Sent and received bytes through a connection can be accessed through Request extensions.

See axum_server::record module for examples.

If accepting connection fails in any one of binded addresses, listening in all binded addresses will be stopped and then an error will be returned.