Crate axum_server

axum-server is a hyper server implementation designed to be used with axum framework.

Modules

service

Service traits.

tls_rustls

Tls implementation using rustls

Structs

Handle

A handle for Server.

Server

HTTP server.

Functions

bind

Create a Server that will bind to provided address.

bind_rustlstls-rustls

Create a RustlsServer that will bind to provided address.